APPS-1332-max-parallel-scans - setting connection queue in client policy

+++ cmd/internal/flags/backup.go
@@ -142,7 +142,7 @@ func (f *Backup) NewFlagSet() *pflag.FlagSet {
 		"How many records will be read on one iteration for continuation backup.\n"+
 			"Affects size if overlap on resuming backup after an error.\n"+
 			"Is used only with --state-file-dst or --continue.")
-	flagSet.Int64Var(&f.MaxParallelScans, "max-parallel-scans",
+	flagSet.IntVar(&f.MaxParallelScans, "max-parallel-scans",
 		5,
 		"Specifies the maximum number of parallel scans per the cluster.\n"+
 			"To avoid no available connections error, you must calculate this parameter wisely.\n"+
