+++ cmd/internal/models/backup.go
@@ -38,7 +38,7 @@ type Backup struct {
 	Continue            string
 	ScanPageSize        int64
 	OutputFilePrefix    string
-	MaxParallelScans    int64
+	MaxParallelScans    int
 }
 
 // ShouldClearTarget check if we should clean target directory.
